Implement FrameMarking header extension support
Frame Marking RTP header extension is used to convey information about
video frames that is critical for error recovery and packet forwarding
in RTP middleboxes or network nodes.

It is most useful when media is encrypted, and essential when the
middlebox or node has no access to the media decryption keys.  It is
also useful for codec-agnostic processing of encrypted or unencrypted
media, while it also supports extensions for codec-specific
information.

https://tools.ietf.org/html/draft-ietf-avtext-framemarking-04